Use all utxos as input and fee based on size
If the first utxo isn't large enough, the trx will fail, so this uses all the utxos as input.
Also tries to get just over 1sat /byte, although the size estimation seems funky, so this can be a bit off. Better to err on the side of too large fees.
These are a bit hacky, but better than what's there atm.